move all ad-hoc random functions into test_util
