+++ test/drive/frontend/pcis_handler.cpp
@@ -9,12 +9,7 @@ TEST(pcis_handler_impl, correct) {
   ST_Queue<PCIS_Write_Req> pcis_write_req_queue(16);
   ST_Queue<SuperCommand> super_command_queue(16);
 
-  PCIS_Write_Req pcis_write_req;
-  pcis_write_req.addr = rand();
-  int int_size_in_bit = sizeof(int) * 8;
-  for (int i = 0; i < PCIS_BUS_WIDTH; i += int_size_in_bit) {
-    pcis_write_req.data(i + int_size_in_bit - 1, i) = rand();
-  }
+  PCIS_Write_Req pcis_write_req = RandomGen::rand_pcis_write_req();
   pcis_write_req_queue.write(pcis_write_req);
 
   RUN_METHOD(1, pcis_handler_impl(&pcis_write_req_queue, &super_command_queue));
